+++ just_audio_background/lib/just_audio_background.dart
@@ -292,6 +292,19 @@ class _JustAudioPlayer extends AudioPlayerPlatform {
       _playerAudioHandler.customSetShuffleOrder(request);
 
   @override
+  Future<SetWebCrossOriginResponse> setWebCrossOrigin(
+      SetWebCrossOriginRequest request) async {
+    _playerAudioHandler.customSetWebCrossOrigin(request);
+    return SetWebCrossOriginResponse();
+  }
+
+  @override
+  Future<SetWebSinkIdResponse> setWebSinkId(SetWebSinkIdRequest request) {
+    _playerAudioHandler.customSetWebSinkId(request);
+    throw SetWebSinkIdResponse();
+  }
+
+  @override
   Future<SeekResponse> seek(SeekRequest request) =>
       _playerAudioHandler.customPlayerSeek(request);
 
@@ -507,10 +520,22 @@ class _PlayerAudioHandler extends BaseAudioHandler
     ));
   }
 
+  Future<SetWebCrossOriginResponse> customSetWebCrossOrigin(
+      SetWebCrossOriginRequest request) async {
+    return await (await _player).setWebCrossOrigin(request);
+  }
+
+  Future<SetWebSinkIdResponse> customSetWebSinkId(
+      SetWebSinkIdRequest request) async {
+    return await (await _player).setWebSinkId(request);
+  }
+
   Future<ConcatenatingInsertAllResponse> customConcatenatingInsertAll(
       ConcatenatingInsertAllRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.insertAll(request.index, request.children);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -521,6 +546,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
       ConcatenatingRemoveRangeRequest request) async {
     final cat = _source!.findCat(request.id)!;
     cat.children.removeRange(request.startIndex, request.endIndex);
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -532,6 +559,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
     final cat = _source!.findCat(request.id)!;
     cat.children
         .insert(request.newIndex, cat.children.removeAt(request.currentIndex));
+    cat.shuffleOrder
+        .replaceRange(0, cat.shuffleOrder.length, request.shuffleOrder);
     _updateShuffleIndices();
     _broadcastStateIfActive();
     _updateQueue();
@@ -582,6 +611,8 @@ class _PlayerAudioHandler extends BaseAudioHandler
       await (await _player).setPreferredPeakBitRate(request);
 
   void _updateQueue() {
+    assert(sequence.every((source) => source.tag is MediaItem),
+        'Error : When using just_audio_background, you should always use a MediaItem as tag when setting an AudioSource. See AudioSource.uri documentation for more information.');
     queue.add(sequence.map((source) => source.tag as MediaItem).toList());
   }