+++ just_audio_platform_interface/lib/just_audio_platform_interface.dart
@@ -230,6 +230,17 @@ abstract class AudioPlayerPlatform {
     throw UnimplementedError(
         "androidEqualizerBandSetGain() has not been implemented.");
   }
+
+  /// Sets the 'crossOrigin' attribute on the web audio element.
+  Future<SetWebCrossOriginResponse> setWebCrossOrigin(
+      SetWebCrossOriginRequest request) {
+    throw UnimplementedError("setWebCrossOrigin() has not been implemented.");
+  }
+
+  /// Sets a specific device output id on the web audio element.
+  Future<SetWebSinkIdResponse> setWebSinkId(SetWebSinkIdRequest request) {
+    throw UnimplementedError("setWebSinkId() has not been implemented.");
+  }
 }
 
 /// A data update communicated from the platform implementation to the Flutter
@@ -1485,3 +1496,21 @@ class AndroidEqualizerMessage extends AudioEffectMessage {
         'parameters': parameters?.toMap(),
       };
 }
+
+class SetWebCrossOriginRequest {
+  final WebCrossOriginMessage? crossOrigin;
+
+  SetWebCrossOriginRequest({required this.crossOrigin});
+}
+
+class SetWebCrossOriginResponse {}
+
+enum WebCrossOriginMessage { anonymous, useCredentials }
+
+class SetWebSinkIdRequest {
+  final String sinkId;
+
+  SetWebSinkIdRequest({required this.sinkId});
+}
+
+class SetWebSinkIdResponse {}