+++ just_audio_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.4.0
+
+* Add setWebSinkId for web (@dganzella).
+
+## 4.3.0
+
+* Add setWebCrossOrigin for CORS on web (@danielwinkler).
+
 ## 4.2.2
 
 * Add setAllowsExternalPlayback on iOS/macOS.